pyproject.toml
[tool.poetry]
name = "afl-technical-test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.7"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.37"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
passlib = "^1.7.4"
alembic = "^1.14.1"
pyjwt = "^2.10.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
